make forces position-based, use tags
The index of an atom inside a molecule is determined by its tag % N,
and the tag // N determines the molecule it is assigned to. This works
only under the assumption that the tags are assigned in order during the
creation of molecules, because the bonds are still used to ensure all
atoms of the own molecule appear as ghost atoms.